Added new quick_sort functions
Each new quick_sort function changes its choice for the pivot
element.  One always selects the first element, one the last
element and one selects the median of the first, middle and
last elements.

I also had to add the some probes for the performance of the
algorithm by performing the number of comparisons each had
to do while partitioning.